adminapi/requirements.txt
django
djangorestframework
pika
psycopg2-binary
python-dotenv
dj-database-url
drf-yasg
model-bakery